Update footprint tool to add exclusion limits.
